extracted helpers from MarketPicker
